+++ lib/src/migration_interface.dart
@@ -1,8 +1,13 @@
+
+import 'package:flutter/foundation.dart';
+
 abstract class Migration{
 
   /// run the migrations
+  @required
   void up();
 
   /// reverse the migrations
+  @required
   void down();
 }